feat: migrating app and welcome components to standalone
