--- src/app/app.module.ts
@@ -1,52 +1,0 @@
-import { MatMenuModule } from '@angular/material/menu';
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { RouterTestingModule } from "@angular/router/testing";
-import { AppRoutingModule } from './app-routing.module';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RouterModule } from '@angular/router';
-import { HttpClientModule } from '@angular/common/http';
-import { MatTableModule } from '@angular/material/table';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { NgHttpCachingModule, NgHttpCachingConfig, NgHttpCachingStrategy } from 'ng-http-caching';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatIconModule } from '@angular/material/icon';
-import { AppComponent } from './app.component';
-import { WelcomeComponent } from './components/welcome/welcome.component';
-import { HeaderComponent } from './components/header/header.component';
-import { MatButtonModule } from '@angular/material/button';
-import { MatDialogModule } from '@angular/material/dialog';
-
-
-const ngHttpCachingConfig: NgHttpCachingConfig = {
-  lifetime: 1000 * 60 * 10, // cache expire after 10 min,
-  allowedMethod: ['GET', 'HEAD'],
-  cacheStrategy: NgHttpCachingStrategy.ALLOW_ALL
-};
-
-@NgModule({
-  declarations: [
-    AppComponent,
-    WelcomeComponent,
-    HeaderComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    RouterModule,
-    RouterTestingModule,
-    HttpClientModule,
-    MatTableModule,
-    MatPaginatorModule,
-    NgHttpCachingModule.forRoot(ngHttpCachingConfig),
-    MatProgressSpinnerModule,
-    MatIconModule,
-    MatMenuModule,
-    MatButtonModule,
-    MatDialogModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
