+++ src/app/components/driver/driver.component.ts
@@ -10,6 +10,7 @@ import { tap } from 'rxjs/operators';
 import { DriversService } from '../../services/drivers.service';
 import { SpinnerService } from '../../services/spinner.service';
 import { BodyTitleComponent } from '../body-title/body-title.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @Component({
   standalone: true,
@@ -17,6 +18,7 @@ import { BodyTitleComponent } from '../body-title/body-title.component';
   templateUrl: './driver.component.html',
   styleUrls: ['./driver.component.scss'],
   imports: [CommonModule, FormsModule, MatFormFieldModule, MatSelectModule, BodyTitleComponent],
+  providers: [DriversService, SpinnerService]
 })
 export class DriverComponent implements OnDestroy {
